Sts read added in protocol

+++ ldr.c
@@ -111,7 +111,7 @@ int16_t LDR_Get_Calculated_Brightness(void){
 }
 
 uint8_t LDR_Automic_Brightness_Sts_Get(void){
-	return LDR.AutomaticBrightness;
+	return (uint8_t)LDR.AutomaticBrightness;
 }
 
 void LDR_Automic_Brightness_On(void){
@@ -122,12 +122,12 @@ void LDR_Automic_Brightness_Off(void){
 	LDR.AutomaticBrightness = FALSE;
 }
 
-void LDR_Manual_Brightness_Set(uint16_t val){
+void LDR_Manual_Brightness_Set(int16_t val){
 	LDR.ManualBrightness = val;
 }
 
 uint16_t LDR_Manual_Brightness_Get(void){
-	return LDR.ManualBrightness;
+	return (uint16_t)LDR.ManualBrightness;
 }
 
 void LDR_Max_Brightness_Set(int16_t val){
